Other minor fixes

+++ tools/nissy_extra.h
@@ -10,5 +10,5 @@ for testing purposes only.
 #include "../src/solvers/tables.h"
 
 size_t gendata_h48_derive(uint8_t, const unsigned char *, unsigned char *);
-int parse_h48_solver(const char *, uint8_t [static 1], uint8_t [static 1]);
+long long parse_h48_solver(const char *, uint8_t [static 1], uint8_t [static 1]);
 long long int nissy_datainfo(uint64_t, const unsigned char *);
